+++ apps/mobile/src/components/common/gestures/index.tsx
@@ -13,6 +13,7 @@ export type GestureAction =
   | 'share'
   | 'hide'
   | 'collapse'
+  | 'collapseThread'
 
 export type GestureData = {
   collapsed?: boolean